Prevent react propTypes warning for CheckboxEditor checkbox (#1378)
* suppress react propTypes warning for checkbox

Make the input element readOnly. It is not attached to a form, and this change makes no functional difference. It does suppress a React propTypes warning that gets printed to console.error.

* suppress react propTypes warning for checkbox

Make the input element readOnly. It is not attached to a form, and this change makes no functional difference. It does suppress a React propTypes warning that gets printed to console.error.

* added tests for CheckboxEditor

* remove test proving console error is not logged